[react] refactor FriendsList, MembersList, and MessageComponent to utilize handleProfileCard prop for profile card functionality
