+++ React/src/components/MembersList.tsx
@@ -8,54 +8,14 @@ import classes from "../styles/application.module.css";
 interface MembersListProps {
   session?: User | null;
   selectedServer?: Server | null;
+  handleProfileCard?: (e: React.MouseEvent, u: User) => void; // Optional prop for handling profile card click
 }
 
 const MembersList: React.FC<MembersListProps> = ({
   session,
   selectedServer,
+  handleProfileCard
 }) => {
-  const [cardVisible, setCardVisible] = useState(false);
-  const [cardPosition, setCardPosition] = useState({ x: 0, y: 0 });
-  const [cardUser, setCardUser] = useState<User | null>(null);
-  const handleProfileCard = (e: React.MouseEvent, u: User) => {
-    e.stopPropagation(); // Stop the event from bubbling up to the document
-    e.preventDefault(); // Prevent the default click
-
-    const viewportWidth = window.innerWidth;
-    const viewportHeight = window.innerHeight;
-    const dropdownWidth = 250; // Approximate width of the dropdown
-    const dropdownHeight = 500; // Approximate height of the dropdown
-
-    let x = e.pageX;
-    let y = e.pageY;
-
-    if (x + dropdownWidth > viewportWidth) {
-      x = viewportWidth - dropdownWidth - 10; // Add some padding
-    }
-    if (y + dropdownHeight > viewportHeight) {
-      y = viewportHeight - dropdownHeight - 10; // Add some padding
-    }
-    setCardPosition({ x, y });
-    setCardUser(u);
-    setCardVisible(true);
-  };
-
-  const handleClickOutsideCard = () => {
-    setCardVisible(false);
-  };
-
-  useEffect(() => {
-    if (cardVisible) {
-      document.addEventListener("click", handleClickOutsideCard);
-    } else {
-      document.removeEventListener("click", handleClickOutsideCard);
-    }
-
-    return () => {
-      document.removeEventListener("click", handleClickOutsideCard);
-    };
-  }, [cardVisible]);
-
   return (
     <>
       <ul className={classes.members_list}>
@@ -87,14 +47,6 @@ const MembersList: React.FC<MembersListProps> = ({
           />
         ))}
       </ul>
-      {/* Profile Card */}
-      {cardVisible && (
-        <UserProfileModal
-          user={cardUser}
-          x={cardPosition.x}
-          y={cardPosition.y}
-        ></UserProfileModal>
-      )}
     </>
   );
 };
